fix for byzantine users
