+++ milestone1/src/main/java/pt/ulisboa/tecnico/milestone1/repository/UserLocationRepository.java
@@ -5,6 +5,6 @@ import pt.ulisboa.tecnico.milestone1.domain.UserLocation;
 import java.util.List;
 
 public interface UserLocationRepository extends JpaRepository<UserLocation, Integer> {
-    UserLocation findUserLocationByUserLocationId_UserIdAndUserLocationId_Epoch(int userId, int epoch);
-    List<UserLocation> findAllByCoordsXAndCoordsYAndUserLocationId_Epoch(int coordsX, int coordsY, int epoch);
+    UserLocation findUserLocationByUserIdAndEpoch(int userId, int epoch);
+    List<UserLocation> findAllByCoordsXAndCoordsYAndEpoch(int coordsX, int coordsY, int epoch);
 }
